refactor : post/domain 폴더로 이동
